Optimization and camera

+++ CoreBase/include/world/Position.hpp
@@ -33,7 +33,7 @@ namespace CubA4::world
 	/*Позиция блока в чанке */
 	using BlockInChunkPos = BasePos<uint8_t>;
 	/*Позиция чанка в мире */
-	using ChunkPos = BasePos<int64_t>;
+	using ChunkPos = BasePos<int32_t>;
 
 	using Layer = uint8_t;
